+++ releasenotes.md
@@ -9,6 +9,7 @@
 - Add json pretty print.
 - `$$atomic_store` and `$$atomic_load` takes an alignment parameter.
 - `$vaarg[^1]` is supported. #3276
+- Improve error message when a keyword is used a block parameter. #3275
 
 ### Stdlib changes
 - Add math::TAU / math::TWO_PI